Khoi Tran - STAT 184 Final Project
This function (s_stats()) scrapes a designated Wikipedia page for data on all 50 U.S. states, and supplements it by merging it with other data sources. The only other file needed for the function to run is the included spreadsheet (state_crime_r.csv), which needs to be in the same project.